goingIND/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==3.2.11
django-cors-headers==3.11.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
idna==3.3
mysqlclient
pysbd==0.3.4
python-decouple==3.5
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.8
Wikipedia-API==0.5.4
